Move `sys_common::rwlock::StaticRWLock` etc. to `sys::unix::rwlock`
